Open Feign clients added

+++ customer-service/src/main/java/ru/antifraud/customer/repository/CustomerRepository.java
@@ -3,5 +3,5 @@ package ru.antifraud.customer.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.antifraud.customer.model.Customer;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 }
